ubi-utils: fix compilation warnings in eb_chain.c when using gcc 4.2.3

+++ ubi-utils/src/eb_chain.c
@@ -221,7 +221,7 @@ eb_chain_print(FILE* stream, struct eb_info *head)
 		else	fprintf(stream, "%-4s", cur->data_crc_ok ? "ok":"ign");
 		fprintf(stream, " %-4d %08x %-8u %-8llu\n", cur->phys_block,
 			cur->phys_addr, be32_to_cpu(cur->vid.data_size),
-			be64_to_cpu(cur->ec.ec));
+			(unsigned long long)be64_to_cpu(cur->ec.ec));
 
 		hist = cur->older;
 		while (hist != NULL) {
@@ -237,7 +237,7 @@ eb_chain_print(FILE* stream, struct eb_info *head)
 			fprintf(stream, " %-4d %08x %-8u %-8llu (*)\n",
 				hist->phys_block, hist->phys_addr,
 				be32_to_cpu(hist->vid.data_size),
-				be64_to_cpu(hist->ec.ec));
+				(unsigned long long)be64_to_cpu(hist->ec.ec));
 
 			hist = hist->older;
 		}
